Bugfix: it was impossible to set the password for the MySQL service when IPOP was enabled.
This bug was a consequence of the way the context file was created: the fresh context template
was used each time a variable substitution was performed, thus erasing any previous partial
variable substitution.
The solution here consists in storing all variables subsitutions in the "Base" cloud and
applying them only once when retrieving the final context. The partially substituted
context is not stored. So the "Base" cloud stores a "_context" (formely called cx_template) and
the "_mapping_vars" to apply to it. A context without variables can be stored, no substitution will
be performed.



git-svn-id: svn://svn.forge.objectweb.org/svnroot/contrail/trunk/conpaas@5134 ef35cc4a-2640-4bbf-98bc-65920db7d8ae